# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Snaptext Developer API",
    "version": "1.0.0",
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Format: Bearer st-<key>"
      }
    },
    "schemas": {
...
      "CreateJobRequest": {
        "type": "object",
        "properties": {
          "pdfUrl": {
            "type": "string",
            "format": "uri",
            "description": "Publicly accessible URL or pre-signed storage URL of the PDF"
          },
          "filename": {
            "type": "string",
            "description": "Original filename of the PDF"
          },
          "fileSize": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "description": "File size in bytes"
          },
          "fileHash": {
            "type": "string",
            "description": "SHA-256 hash of the file to prevent duplicate runs"
          },
          "ocrModelId": {
            "type": "string",
            "nullable": true,
            "description": "OCR Model Name (e.g., 'Flux', 'Spark', 'Zenith'), Model ID (e.g., 'google/gemini-3.1-flash-lite-preview'), or model UUID"
          },
          "jsonSchema": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object",
                "additionalProperties": {}
              }
            ],
            "nullable": true,
            "description": "Optional JSON Schema for custom structured extraction layout"
          }
        },
        "required": [
          "pdfUrl",
          "filename",
          "fileSize",
          "fileHash"
        ]
      },
      "CreateUploadUrlRequest": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "description": "Original name of the file to be uploaded"
          },
          "fileSize": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "description": "File size in bytes (max 20MB)"
          },
          "fileHash": {
            "type": "string",
            "description": "SHA-256 hash of the file payload"
          },
          "ocrModelId": {
            "type": "string",
            "nullable": true,
            "description": "OCR Model Name (e.g., 'Flux', 'Spark', 'Zenith'), Model ID (e.g., 'google/gemini-3.1-flash-lite-preview'), or model UUID"
          }
        },
        "required": [
          "filename",
          "fileSize",
          "fileHash"
        ]
      },
      "CreateUploadUrlResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Ephemeral client upload token for storage"
          },
          "pathname": {
            "type": "string",
            "description": "Storage pathname destination"
          },
          "uploadUrl": {
            "type": "string",
            "format": "uri",
            "description": "Direct storage upload target URL"
          },
          "deduplicated": {
            "type": "boolean",
            "description": "Indicates if an identical job is already processing"
          },
          "jobId": {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the existing job if deduplicated"
          },
          "runId": {
            "type": "string",
            "description": "Workflow execution run ID if deduplicated"
          },
          "pdfUrl": {
            "type": "string",
            "format": "uri",
            "description": "Source URL of the existing job if deduplicated"
          }
        }
      },
...
      "JobListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "runId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "fileSize": {
            "type": "number"
          },
          "pdfUrl": {
            "type": "string"
          },
          "totalPages": {
            "type": "number",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "runId",
          "status",
          "filename",
          "fileSize",
          "pdfUrl",
          "totalPages",
          "error",
          "createdAt",
          "updatedAt"
        ]
      },
      "JobListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/JobListItem"
        }
      },
      "JobSubmissionResponse": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the submitted job"
          },
          "runId": {
            "type": "string",
            "description": "Underlying workflow execution run ID"
          },
          "pdfUrl": {
            "type": "string",
            "format": "uri",
            "description": "Validated storage/source URL of the PDF"
          },
          "status": {
            "type": "string",
            "description": "Status of the job, defaults to pending"
          },
          "deduplicated": {
            "type": "boolean",
            "description": "Indicates if an active job was reused"
          }
        },
        "required": [
          "jobId",
          "runId",
          "pdfUrl",
          "status"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/v1/jobs": {
      "get": {
        "operationId": "get-v1-jobs",
        "summary": "List Jobs",
        "description": "Retrieve a paginated list of recent document extraction jobs submitted by the user.",
        "tags": [
          "V1"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Number of items to return",
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "description": "Number of items to return"
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Offset for pagination",
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            },
            "description": "Offset for pagination"
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-v1-jobs",
        "summary": "Submit OCR Job",
        "description": "Initiates a new structured document extraction job. Accepts a direct URL or a pre-uploaded storage URL.",
        "tags": [
          "V1"
        ],
        "parameters": [],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmissionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/v1/jobs/upload-url": {
      "post": {
        "operationId": "post-v1-jobs-upload-url",
        "summary": "Generate Upload Token",
        "description": "Generates a secure client token to upload PDF binaries directly to secure storage.",
        "tags": [
          "V1"
        ],
        "parameters": [],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadUrlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUploadUrlResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
...
}